bridge Swift Interface to C++ Interface

+++ apple/objc/core/WCTTag.h
@@ -24,5 +24,5 @@
 
 #import <stdint.h>
 
-typedef int32_t WCTTag;
+typedef long WCTTag;
 static WCTTag const WCTInvalidTag = 0;
